+++ src/lib/wait_for_remote_peer.ts
@@ -22,7 +22,8 @@ interface WakuGossipSubProtocol extends GossipSub {
 
 /**
  * Wait for a remote peer to be ready given the passed protocols.
- * Useful when using the [[CreateOptions.bootstrap]] with [[createWaku]].
+ * Must be used after attempting to connect to nodes, using {@link Waku.dial} or
+ * a bootstrap method with {@link Waku.constructor}.
  *
  * If the passed protocols is a GossipSub protocol, then it resolves only once
  * a peer is in a mesh, to help ensure that other peers will send and receive